requirements.txt
Django==5.0
djangorestframework==3.14.0 
aima3==1.0.11